+++ libs/cardano-ledger-core/src/Cardano/Ledger/Address.hs
@@ -188,7 +188,7 @@ data AccountAddress = AccountAddress
 
 newtype AccountId = AccountId {unAccountId :: Credential Staking}
   deriving (Show, Eq, Generic, Ord)
-  deriving newtype (NFData, NoThunks, ToJSON, FromJSON, EncCBOR, DecCBOR)
+  deriving newtype (NFData, NoThunks, ToJSON, ToJSONKey, FromJSON, FromJSONKey, EncCBOR, DecCBOR)
 
 -- | Deprecated pattern synonym for backward compatibility
 pattern RewardAccount :: Network -> Credential Staking -> AccountAddress