+++ eras/dijkstra/impl/CHANGELOG.md
@@ -111,7 +111,13 @@
   - `DijkstraUtxowPredFailure`
 * Remove `NoThunks` instance for `DijkstraContextError`
 * Make `DijkstraContextError` constructors lazy
-* Add `ToJSON` and `FromJSON` instances for `DijkstraNativeScript era`
+* Add `ToJSON` and `FromJSON` instances for
+  - `DijkstraNativeScript era`
+  - `AccountBalanceInterval era`
+* Add `FromJSON` instance for
+  - `DijkstraScript era`
+  - `DijkstraDelegCert`
+  - `DijkstraTxCert era`
 * Export `dijkstraBasedEraNativeScriptToJSON` and `dijkstraBasedEraNativeScriptJSONParser` from `Cardano.Ledger.Dijkstra.Scripts`
 * Remove DRep requirement for reward withdrawals
   - Remove `WdrlNotDelegatedToDRep` constructor from `EntitiesPredFailure`